Refactor based on Ed's feedback to create primordial store
