+++ packages/host/app/components/operator-mode/code-submode.gts
@@ -242,13 +242,11 @@ export default class CodeSubmode extends Component<Signature> {
   }
 
   private get card() {
-    if (
-      this.cardResource.card &&
-      this.codePath?.href.replace(/\.json$/, '') === this.cardResource.url
-    ) {
-      return this.cardResource.card;
-    }
-    return undefined;
+    return this.cardResource.card;
+  }
+
+  private get cardError() {
+    return this.cardResource.cardError;
   }
 
   private backgroundURLStyle(backgroundURL: string | null) {
@@ -318,11 +316,6 @@ export default class CodeSubmode extends Component<Signature> {
 
   private get isEmptyFile() {
     return this.readyFile.content.match(/^\s*$/);
-  }
-
-  @cached
-  get cardError() {
-    return this.cardResource.cardError;
   }
 
   @cached
@@ -450,21 +443,21 @@ export default class CodeSubmode extends Component<Signature> {
     oldCard: CardDef | undefined,
     newCard: CardDef | undefined,
   ) => {
-    if (oldCard) {
-      this.cardResource.api.unsubscribeFromChanges(oldCard, this.onCardChange);
-    }
-    if (newCard) {
-      this.cardResource.api.subscribeToChanges(newCard, this.onCardChange);
+    // this handles the scenario for the initial load, as well as unloading a
+    // card that went into an error state or was deleted
+    if (oldCard !== newCard) {
+      if (oldCard) {
+        this.cardResource.api.unsubscribeFromChanges(
+          oldCard,
+          this.onCardChange,
+        );
+      }
+      if (newCard) {
+        this.cardResource.api.subscribeToChanges(newCard, this.onCardChange);
+      }
     }
     this.#currentCard = newCard;
   };
-
-  private get loadedCard() {
-    if (!this.card) {
-      throw new Error(`bug: card ${this.codePath} is not loaded`);
-    }
-    return this.card;
-  }
 
   private get declarations() {
     return this.moduleContentsResource?.declarations;
@@ -1068,7 +1061,7 @@ export default class CodeSubmode extends Component<Signature> {
                     </Accordion>
                   {{else if this.card}}
                     <CardPreviewPanel
-                      @card={{this.loadedCard}}
+                      @card={{this.card}}
                       @realmURL={{this.realmURL}}
                       @format={{this.previewFormat}}
                       @setFormat={{this.setPreviewFormat}}
