+++ packages/host/app/resources/card-resource.ts
@@ -4,51 +4,24 @@
 
 import { registerDestructor } from '@ember/destroyable';
 import { getOwner } from '@ember/owner';
-import { service } from '@ember/service';
 import { tracked } from '@glimmer/tracking';
 
 import { restartableTask } from 'ember-concurrency';
-import { task } from 'ember-concurrency';
+
 import { Resource } from 'ember-resources';
 
-import status from 'statuses';
-
-import {
-  isSingleCardDocument,
-  apiFor,
-  isCardInstance,
-  type SingleCardDocument,
-  type LooseSingleCardDocument,
-} from '@cardstack/runtime-common';
-
-import type MessageService from '@cardstack/host/services/message-service';
+import { type LooseSingleCardDocument } from '@cardstack/runtime-common';
 
 import type { CardDef } from 'https://cardstack.com/base/card-api';
-
 import type * as CardAPI from 'https://cardstack.com/base/card-api';
 
+import { type CardError } from '../services/store';
+
+import { asURL } from '../services/store';
+
 import type CardService from '../services/card-service';
-import type LoaderService from '../services/loader-service';
-import type RealmService from '../services/realm';
-import type RealmSubscriptionService from '../services/realm-subscription';
 
-interface CardErrors {
-  errors: {
-    id?: string; // 404 errors won't necessarily have an id
-    status: number;
-    title: string;
-    message: string;
-    realm: string | undefined;
-    meta: {
-      lastKnownGoodHtml: string | null;
-      cardTitle: string | null;
-      scopedCssUrls: string[];
-      stack: string | null;
-    };
-  }[];
-}
-
-export type CardError = CardErrors['errors'][0];
+import type StoreService from '../services/store';
 
 interface Args {
   named: {
@@ -56,12 +29,9 @@ interface Args {
     // different instances don't result in re-running the resource
     urlOrDoc: string | LooseSingleCardDocument | undefined;
     isLive: boolean;
-    relativeTo?: URL; // used for new cards
     // this is not always constructed within a container so we pass in our services
+    storeService: StoreService;
     cardService: CardService;
-    realmSubscriptionService: RealmSubscriptionService;
-    messageService: MessageService;
-    resetLoader: () => void;
     onCardInstanceChange?: (
       oldCard: CardDef | undefined,
       newCard: CardDef | undefined,
@@ -70,198 +40,98 @@ interface Args {
 }
 
 export class CardResource extends Resource<Args> {
-  url: string | undefined;
-  @tracked loaded: Promise<void> | undefined;
-  @tracked cardError: CardError | undefined;
-  @service private declare realm: RealmService;
+  // we use a separate tracked property for the card instead of directly
+  // punching thru to the store, since using a TrackedMap in the store's
+  // IdentityContext would result in a Resource.modify() cycle as the
+  // IdentityContext is mutated as part of loading the card.
   @tracked private _card: CardDef | undefined;
-  @tracked private _api: typeof CardAPI | undefined;
-  @tracked private staleCard: CardDef | undefined;
-  private relativeTo: URL | undefined;
-  private isLive = false;
-  private declare realmSubscription: RealmSubscriptionService;
+  // and just being symmetric with the card error as well
+  @tracked private _error: CardError | undefined;
+  @tracked private _isLoaded = false;
+  @tracked private _loaded: Promise<void> | undefined;
+  private declare store: StoreService;
   private declare cardService: CardService;
-  private declare messageService: MessageService;
-  private declare loaderService: LoaderService;
-  private declare resetLoader: () => void;
-  private onCardInstanceChange?: (
+  #url: string | undefined;
+  #isLive = false;
+  #api: typeof CardAPI | undefined;
+
+  onCardInstanceChange?: (
     oldCard: CardDef | undefined,
     newCard: CardDef | undefined,
   ) => void;
 
   modify(_positional: never[], named: Args['named']) {
-    let {
-      urlOrDoc,
-      isLive,
-      onCardInstanceChange,
-      messageService,
-      cardService,
-      realmSubscriptionService,
-      resetLoader,
-      relativeTo,
-    } = named;
-    this.relativeTo = relativeTo;
-    this.messageService = messageService;
+    let { urlOrDoc, isLive, onCardInstanceChange, storeService, cardService } =
+      named;
+    this.store = storeService;
     this.cardService = cardService;
-    this.realmSubscription = realmSubscriptionService;
-    this.url = urlOrDoc ? asURL(urlOrDoc) : undefined;
+    this.#url = urlOrDoc ? asURL(urlOrDoc) : undefined;
+    this.#isLive = isLive;
     this.onCardInstanceChange = onCardInstanceChange;
-    this.cardError = undefined;
-    this.resetLoader = resetLoader;
-    this.isLive = isLive;
+
     if (urlOrDoc) {
-      this.loaded = this.loadModel.perform(urlOrDoc);
+      this._loaded = this.load.perform(urlOrDoc);
     }
 
     registerDestructor(this, () => {
-      this.realmSubscription.unloadResource(this);
+      this.store.unloadResource(this);
     });
   }
 
+  get isLive() {
+    return this.#isLive;
+  }
+
   get card() {
-    if (this.loadModel.isRunning) {
-      return this.staleCard;
-    }
     return this._card;
   }
 
+  get url() {
+    return this.#url;
+  }
+
+  get cardError() {
+    return this._error;
+  }
+
+  get isLoaded() {
+    return this._isLoaded;
+  }
+
+  // This is deprecated. consumers of this resource need to be reactive such
+  // that they can deal with a resource that doesn't have a card yet.
+  get loaded() {
+    return this._loaded;
+  }
+
+  private load = restartableTask(
+    async (urlOrDoc: string | LooseSingleCardDocument) => {
+      this.#api = await this.cardService.getAPI();
+      let { url, card, error } = await this.store.createSubscriber({
+        resource: this,
+        urlOrDoc,
+        setCard: (card) => {
+          if (card !== this.card) {
+            this._card = card;
+          }
+        },
+        setCardError: (error) => (this._error = error),
+      });
+      this.#url = url;
+      this._card = card;
+      this._error = error;
+      this._isLoaded = true;
+    },
+  );
+
+  // TODO refactor this out
   get api() {
-    if (!this._api) {
+    if (!this.#api) {
       throw new Error(
         `API hasn't been loaded yet in CardResource--await this.loaded()`,
       );
     }
-    return this._api;
-  }
-
-  loadModel = restartableTask(
-    async (urlOrDoc: string | LooseSingleCardDocument) => {
-      let cardOrError = await this.getCard(urlOrDoc);
-      await this.updateCardInstance(cardOrError);
-      if (!isCardInstance(cardOrError)) {
-        console.warn(`cannot load card ${cardOrError.id}`, cardOrError);
-      }
-    },
-  );
-
-  private async getCard(urlOrDoc: string | LooseSingleCardDocument) {
-    let url = asURL(urlOrDoc);
-    let identityContext = this.realmSubscription.identityContext;
-    try {
-      if (!url) {
-        // this is a new card so instantiate it and save it
-        let doc = urlOrDoc as LooseSingleCardDocument;
-        let newCard = await this.cardService.createFromSerialized(
-          doc.data,
-          doc,
-          this.relativeTo,
-          {
-            identityContext,
-          },
-        );
-        await this.cardService.saveModel(newCard);
-        if (identityContext) {
-          identityContext.set(newCard.id, newCard);
-        }
-        return newCard;
-      }
-
-      // createFromSerialized would also do this de-duplication, but we want to
-      // also avoid the fetchJSON when we already have the stable card.
-      let existingCard = identityContext?.get(url);
-      if (existingCard) {
-        return existingCard;
-      }
-      let doc = (typeof urlOrDoc !== 'string' ? urlOrDoc : undefined) as
-        | SingleCardDocument
-        | undefined;
-      if (!doc) {
-        let json = await this.cardService.fetchJSON(url);
-        if (!isSingleCardDocument(json)) {
-          throw new Error(
-            `bug: server returned a non card document for ${url}:
-        ${JSON.stringify(json, null, 2)}`,
-          );
-        }
-        doc = json;
-      }
-      let card = await this.cardService.createFromSerialized(
-        doc.data,
-        doc,
-        new URL(doc.data.id),
-        {
-          identityContext,
-        },
-      );
-      if (identityContext && !identityContext.get(url)) {
-        identityContext.set(url, card);
-      }
-      return card;
-    } catch (error: any) {
-      if (url && identityContext) {
-        identityContext.set(url, undefined);
-      }
-      let errorResponse = processCardError(
-        url ? new URL(url) : undefined,
-        error,
-      );
-      return errorResponse.errors[0];
-    }
-  }
-
-  reload = task(async (card: CardDef) => {
-    try {
-      await this.cardService.reloadCard(card);
-      this.setCardOrError(card);
-    } catch (err: any) {
-      if (err.status !== 404) {
-        this.realmSubscription.identityContext.set(card.id, undefined);
-        let errorResponse = processCardError(new URL(card.id), err);
-        this.setCardOrError(errorResponse.errors[0]);
-        return;
-      }
-      // in this case the document was invalidated in the index because the
-      // file was deleted
-      this.clearCardInstance();
-      return;
-    }
-  });
-
-  async updateCardInstance(maybeCard: CardDef | CardError) {
-    let instance: CardDef | undefined;
-    if (isCardInstance(maybeCard)) {
-      instance = maybeCard;
-      this.url = maybeCard.id;
-      this._api = await apiFor(maybeCard);
-    }
-    if (this.onCardInstanceChange) {
-      this.onCardInstanceChange(this._card, instance);
-    }
-    if (this.isLive) {
-      this.realmSubscription.subscribeFor(this);
-    }
-    this.setCardOrError(maybeCard);
-  }
-
-  private setCardOrError(cardOrError: CardDef | CardError) {
-    if (isCardInstance(cardOrError)) {
-      this._card = cardOrError;
-      this.staleCard = cardOrError;
-      this.cardError = undefined;
-    } else {
-      this.cardError = cardOrError;
-      this._card = undefined;
-      this.staleCard = undefined;
-    }
-  }
-
-  private clearCardInstance() {
-    if (this.onCardInstanceChange) {
-      this.onCardInstanceChange(this._card, undefined);
-    }
-    this._api = undefined;
-    this._card = undefined;
-    this.staleCard = undefined;
+    return this.#api;
   }
 }
 
@@ -271,92 +141,27 @@ export function getCard(
   opts?: {
     relativeTo?: URL; // used for new cards
     isLive?: () => boolean;
+    // TODO refactor this out
     onCardInstanceChange?: () => (
       oldCard: CardDef | undefined,
       newCard: CardDef | undefined,
     ) => void;
   },
 ) {
-  let loaderService = (getOwner(parent) as any).lookup(
-    'service:loader-service',
-  ) as LoaderService;
   return CardResource.from(parent, () => ({
     named: {
       urlOrDoc: urlOrDoc(),
       isLive: opts?.isLive ? opts.isLive() : true,
-      relativeTo: opts?.relativeTo,
       onCardInstanceChange: opts?.onCardInstanceChange
         ? opts.onCardInstanceChange()
         : undefined,
-      resetLoader: loaderService.reset.bind(loaderService),
-      realmSubscriptionService: (getOwner(parent) as any).lookup(
-        'service:realm-subscription',
-      ) as RealmSubscriptionService,
-      messageService: (getOwner(parent) as any).lookup(
-        'service:message-service',
-      ) as MessageService,
+      storeService: (getOwner(parent) as any).lookup(
+        'service:store',
+      ) as StoreService,
+      // TODO refactor this out
       cardService: (getOwner(parent) as any).lookup(
         'service:card-service',
       ) as CardService,
     },
   }));
-}
-
-function processCardError(url: URL | undefined, error: any): CardErrors {
-  try {
-    let errorResponse = JSON.parse(error.responseText) as CardErrors;
-    return errorResponse;
-  } catch (parseError) {
-    switch (error.status) {
-      // tailor HTTP responses as necessary for better user feedback
-      case 404:
-        return {
-          errors: [
-            {
-              id: url?.href,
-              status: 404,
-              title: 'Card Not Found',
-              message: `The card ${url?.href} does not exist`,
-              realm: error.responseHeaders?.get('X-Boxel-Realm-Url'),
-              meta: {
-                lastKnownGoodHtml: null,
-                scopedCssUrls: [],
-                stack: null,
-                cardTitle: null,
-              },
-            },
-          ],
-        };
-      default:
-        return {
-          errors: [
-            {
-              id: url?.href,
-              status: error.status ?? 500,
-              title: error.status
-                ? status.message[error.status]
-                : error.message,
-              message: error.status
-                ? `Received HTTP ${error.status} from server ${
-                    error.responseText ?? ''
-                  }`.trim()
-                : `${error.message}: ${error.stack}`,
-              realm: error.responseHeaders?.get('X-Boxel-Realm-Url'),
-              meta: {
-                lastKnownGoodHtml: null,
-                scopedCssUrls: [],
-                stack: null,
-                cardTitle: null,
-              },
-            },
-          ],
-        };
-    }
-  }
-}
-
-export function asURL(urlOrDoc: string | LooseSingleCardDocument) {
-  return typeof urlOrDoc === 'string'
-    ? urlOrDoc.replace(/\.json$/, '')
-    : urlOrDoc.data.id;
 }
